fix unnecessary cast warnings
git-svn-id: https://svn.apache.org/repos/asf/poi/trunk@1715626 13f79535-47bb-0310-9956-ffa450edef68

+++ src/scratchpad/src/org/apache/poi/hwpf/model/Sttb.java
@@ -214,7 +214,7 @@ public class Sttb
 
             if ( _fExtend )
             {
-                LittleEndian.putUShort( buffer, offset, (int) entry.length() );
+                LittleEndian.putUShort( buffer, offset, entry.length() );
                 offset += LittleEndian.SHORT_SIZE;
 
                 StringUtil.putUnicodeLE( entry, buffer, offset );
